more smatrix restructures
