+++ src/dmatrix/mod.rs
@@ -13,8 +13,8 @@
 use crate::common::{
     Dimension, EvalInto, Field, Float, FloatVectorOps, IndexValue, LazyExpr, One, VectorOps, Zero,
 };
-use crate::expression::binary_expr::{BinOp, BinopExpr};
-use crate::expression::unary_expr::{UnaryExpr, UnaryOp};
+use crate::expression::binary_expr::BinOp;
+use crate::expression::unary_expr::UnaryOp;
 //}}}
 //{{{ std imports
 use serde::{Deserialize, Serialize};
