+++ src/smatrix/blaslapack/mod.rs
@@ -11,11 +11,17 @@
 //}}}
 //--------------------------------------------------------------------------------------------------
 
+mod eig;
+mod lu;
+mod matmul;
+mod qr;
+mod schur;
+mod solve;
+mod symeig;
 
-pub mod eig;
-pub mod lu;
-pub mod matmul;
-pub mod qr;
-pub mod schur;
-pub mod solve;
-pub mod symeig;
+pub use eig::{Error as SEigError, Return as SEigReturn};
+pub use lu::{Error as SLuError, Return as SLuReturn};
+pub use qr::{Error as SQrError, Return as SQrReturn};
+pub use schur::{Error as SSchurError, Return as SSchurReturn};
+pub use solve::Error as SSolveError;
+pub use symeig::{Error as SSymEigError, Return as SSymEigReturn};
