+++ tests/matrix_ops.rs
@@ -6,8 +6,7 @@ mod smatrix_tests
 
     use approx::assert_relative_eq;
     use topohedral_linalg::{
-        DMatrix, SMatrix, DQrReturn,
-        DVector, VecType, Complex, MatMul, MatrixOps, Shape,
+        Complex, DVector, MatMul, MatrixOps, SMatrix, SQrReturn, SSchurReturn, Shape, VecType,
     };
 
     //{{{ collection: eig tests
@@ -438,7 +437,7 @@ mod smatrix_tests
             12.0, -51.0, 4.0, 6.0, 167.0, -68.0, -4.0, 24.0, -41.0,
         ]);
 
-        let DQrReturn{ q, r } = a.qr().unwrap();
+        let SQrReturn { q, r } = a.qr().unwrap();
 
         // Verify Q*R = A
         let reconstructed: SMatrix<f64, 3, 3> = q.matmul(&r);
@@ -475,7 +474,7 @@ mod smatrix_tests
         let a =
             SMatrix::<f64, 3, 3>::from_row_slice(&[1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0, 9.0]);
 
-        let schur::Return { q, t } = a.schur().unwrap();
+        let SSchurReturn { q, t } = a.schur().unwrap();
 
         // Verify Q*T*Q^T = A
         let q_t = q.matmul(&t);
